feat(services): adds unanchored property to marathon errors
This adds a new `unanchored` property to errors which enable the `ErrorMessageUtil` to remove the path from the displayed message.

This change will allow the ErrorAlert to de-duplicate these messages.

Example:
```
cmd: You must specify a command, an argument or a container
args: You must specify a command, an argument or a container
container.docker.image: You must specify a command, an argument or a container
```
becomes
```
You must specify a command, an argument or a container
```

Closes DCOS-14753